feat: enhance Profile page with admin badge and improve text formatting
- Added conditional rendering for admin badge on Profile page.
- Improved text formatting for better readability in various sections.
- Updated delete account confirmation text for clarity.

chore: add Redis service to Docker Compose configuration

- Introduced Redis service with health checks and volume management.

fix: log response in deactivateRepoActivity and fetch recent user logs in admin analytics

- Added logging for response in deactivateRepoActivity function.
- Implemented fetching of the three latest user logs in fetchAdminAnalytics.
